requirements.txt
comet-ml==2.0.16
h5py==2.9.0
ipython==7.9.0
jupyter==1.0.0
Pillow==6.2.1
tensorflow==2.0.0
tensorflow-addons==0.6.0
pyyaml==5.1.2
scikit-image==0.16.2
tqdm==4.39.0

pytest==5.2.2
